pyproject.toml
[project]
name = "strava-mcp"
version = "0.1.0"
description = "Standalone MCP server for read-only personal Strava inspection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
]

[project.scripts]
strava-mcp = "strava_mcp.main:main"
strava = "strava_mcp.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strava_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-q --tb=short --no-header"
pythonpath = ["src"]
